[mem_cache][1/N] refactor: split allocator.py into allocator/ subpackage
For allocator/base.py and allocator/token.py,
`git blame -C -C -C` recovers the full pre-refactor ancestry.
